ALL_DATA is the fully merged metrics dataframe
Okay I'm uploading manually because my terminal keeps stalling out when I push to git. 
The ALL_DATA csv file is the dataframe where I've merged all of our isolated metrics.
The MAIN_CODE notebook file is just all the code where I did all of that (along with all the extra cleaning and what not to get it to work).